Move container files
