+++ src/ts/containers/Binder.ts
@@ -1,7 +1,7 @@
 import * as ReactRedux from 'react-redux';
 
-import {Binder} from './components/Binder';
-import {BinderState} from './reducer';
+import {Binder} from '../components/Binder';
+import {BinderState} from '../reducer';
 
 const mapStateToProps = (state: BinderState) => state
