+++ src/ts/containers/Form.ts
@@ -1,7 +1,7 @@
 import * as ReactRedux from 'react-redux';
 
-import {Form} from './components/Form';
-import {FormState} from './reducer';
+import {Form} from '../components/Form';
+import {FormState} from '../reducer';
 
 const mapStateToProps = (state: FormState) => state
